Breaking out CSS and JS function from main CssSweet class Fixing warning from missing properties
