+++ core/components/csssweet/src/v2/Event/Event.php
@@ -29,7 +29,7 @@ abstract class Event
         return $this->modx->getOption($key, $this->sp, $default, $skipEmpty);
     }
 
-    protected function debug(string $msg, mixed $data = null): void
+    protected function debugMsg(string $msg, mixed $data = null): void
     {
         if (!$this->debug) {
             return;
