+++ core/components/csssweet/src/v2/Snippet/Extract.php
@@ -1,15 +1,17 @@
 <?php
 namespace CssSweet\v2\Snippet;
 
+use CssSweet\v2\Traits\CSS;
 use CssSweet\v2\Traits\Modifier;
 
 class Extract extends Snippet
 {
+    use CSS;
     use Modifier;
     public function modify($input, $options)
     {
         // Set color class
-        $cc = $this->cs->getColorClass($input);
+        $cc = $this->getColorClass($input);
         if (!$cc['color']) {
             return '';
         }
